+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Bugfixes
 - Fixed creative mode item hiding running even when the server has the mod.
+- Fixed Advanced Barrier blocks not respecting config settings.
+- Fixed Advanced Barrier blocks not turning visible when a barrier is held.